Рефакторинг кода. Первый поход
