+++ src/main/kotlin/ru/marslab/ide/ride/agent/testing/LanguageTestingAgent.kt
@@ -1,4 +1,6 @@
-package ru.marslab.ide.ride.testing
+package ru.marslab.ide.ride.agent.testing
+
+import ru.marslab.ide.ride.testing.GeneratedTest
 
 /**
  * Языковой агент: генерирует тесты для конкретного языка (Kotlin/Java/Dart).
