+++ src/main/kotlin/ru/marslab/ide/ride/service/ChatService.kt
@@ -39,7 +39,6 @@ import ru.marslab.ide.ride.agent.a2a.MessagePayload
 import ru.marslab.ide.ride.agent.a2a.MessageBusProvider
 import ru.marslab.ide.ride.testing.TestRunResult
 import ru.marslab.ide.ride.testing.TestingAgentOrchestrator
-import ru.marslab.ide.ride.testing.TestingAgentOrchestratorImpl
 import java.time.Instant
 
 /**
@@ -684,8 +683,7 @@ class ChatService {
 
                     val relativePath = atToken.removePrefix("@")
 
-                    // Реальный оркестратор (MVP)
-                    val orchestrator: TestingAgentOrchestrator = TestingAgentOrchestratorImpl()
+                    val orchestrator = TestingAgentOrchestrator()
                     val result: TestRunResult = runCatching { orchestrator.generateAndRun(relativePath) }
                         .getOrElse { ex ->
                             withContext(Dispatchers.EDT) {
