+++ src/main/kotlin/ru/marslab/ide/ride/testing/TestPersister.kt
@@ -1,6 +1,5 @@
 package ru.marslab.ide.ride.testing
 
-import java.nio.file.Path
 /**
  * Сохранение сгенерированных тестов в структуру проекта.
  */
@@ -15,5 +14,6 @@ interface TestPersister {
      * Сохраняет сгенерированные тесты с учётом исходного файла (relativePath внутри workspace).
      * По умолчанию делегирует на persist(). Реализации могут учитывать путь для выбора папки/импорта.
      */
-    suspend fun persistForSource(sourceRelativePath: String, tests: List<GeneratedTest>): List<java.nio.file.Path> = persist(tests)
+    suspend fun persistForSource(sourceRelativePath: String, tests: List<GeneratedTest>): List<java.nio.file.Path> =
+        persist(tests)
 }
