+++ src/main/kotlin/ru/marslab/ide/ride/testing/TestingAgentOrchestrator.kt
@@ -1,12 +1,110 @@
 package ru.marslab.ide.ride.testing
 
+import com.intellij.openapi.application.ApplicationManager
+import com.intellij.openapi.diagnostic.Logger
+import com.intellij.openapi.fileEditor.FileEditorManager
+import com.intellij.openapi.project.Project
+import com.intellij.openapi.project.ProjectManager
+import com.intellij.openapi.vfs.LocalFileSystem
+import ru.marslab.ide.ride.agent.testing.DartTestingAgent
+import ru.marslab.ide.ride.agent.testing.JavaTestingAgent
+import ru.marslab.ide.ride.agent.testing.KotlinTestingAgent
+import ru.marslab.ide.ride.agent.testing.LanguageTestingAgent
+import java.nio.file.Files
+
 /**
- * Оркестратор генерации и запуска unit-тестов для выбранного файла исходников.
+ * Реализация оркестратора: анализ → генерация → сохранение → запуск.
  */
-interface TestingAgentOrchestrator {
-    /**
-     * Полный цикл: анализ файла, генерация тестов, сохранение и запуск.
-     * @param filePath относительный путь внутри workspace (начиная от корня проекта)
-     */
-    suspend fun generateAndRun(filePath: String): TestRunResult
+class TestingAgentOrchestrator(
+    private val structureProvider: ProjectStructureProvider = A2AProjectStructureProvider(),
+    private val testRunner: TestRunner = TerminalBasedTestRunner(A2AProjectStructureProvider()),
+    private val persister: TestPersister = FileSystemTestPersister(A2AProjectStructureProvider()),
+    private val detectors: List<LanguageTestingAgent> = listOf(
+        KotlinTestingAgent(), JavaTestingAgent(),
+        DartTestingAgent()
+    ),
+    private val frameworkDetector: TestFrameworkDetector = JvmTestFrameworkDetector(A2AProjectStructureProvider())
+) {
+    private val logger = Logger.getInstance(TestingAgentOrchestrator::class.java)
+
+    suspend fun generateAndRun(filePath: String): TestRunResult {
+        logger.debug("TestingOrchestrator: start generateAndRun filePath='$filePath'")
+        val t0 = System.currentTimeMillis()
+        val structure = structureProvider.getProjectStructure().also {
+            logger.debug("TestingOrchestrator: structure buildSystem=${it.buildSystem} root='${it.root}' testDirs=${it.testSourceDirs}")
+        }
+
+        val abs = structure.root.resolve(filePath).normalize()
+        require(Files.exists(abs)) { "Файл не найден: $filePath" }
+
+        val langAgent = detectors.firstOrNull { it.supports(filePath) }
+            ?: return TestRunResult(
+                success = false,
+                passed = 0,
+                failed = 0,
+                skipped = 0,
+                durationMs = 0,
+                reportText = "Не поддерживаемый тип файла для генерации тестов",
+                errors = listOf("Unsupported language for $filePath")
+            )
+        logger.debug("TestingOrchestrator: selected languageAgent='${langAgent::class.simpleName}'")
+
+        val source = Files.readString(abs)
+        logger.debug("TestingOrchestrator: source loaded, length=${source.length}")
+
+        // TODO: на этапе анализа учитывать приватные элементы, но генерировать тесты на публичный API
+        val tGen0 = System.currentTimeMillis()
+        val tests = if (structure.buildSystem == BuildSystem.DART && langAgent is DartTestingAgent) {
+            val enriched = DartTestingAgent.enrichWithDartMetadata(structure.root, filePath, source)
+            langAgent.generate(enriched)
+        } else {
+            langAgent.generate(source)
+        }
+        logger.debug("TestingOrchestrator: generated tests count=${tests.size} in ${System.currentTimeMillis() - tGen0}ms")
+
+        val tSave0 = System.currentTimeMillis()
+        val savedPaths = persister.persistForSource(filePath, tests)
+        logger.debug("TestingOrchestrator: saved tests to ${savedPaths.joinToString()} in ${System.currentTimeMillis() - tSave0}ms")
+
+        // Откроем первый созданный тестовый файл в IDE
+        val project: Project? = ProjectManager.getInstance().openProjects.firstOrNull()
+        val firstTest = savedPaths.firstOrNull()
+        if (project != null && firstTest != null) {
+            ApplicationManager.getApplication().invokeLater {
+                val vFile = LocalFileSystem.getInstance().refreshAndFindFileByPath(firstTest.toString())
+                if (vFile != null) {
+                    FileEditorManager.getInstance(project).openFile(vFile, true)
+                }
+            }
+        }
+
+        // Попробуем запустить только сгенерированный класс
+        val scope = when (structure.buildSystem) {
+            BuildSystem.DART -> {
+                // для Dart передаем путь к тест-файлу, чтобы запустить конкретный тест
+                savedPaths.firstOrNull()?.let { structure.root.relativize(it).toString() }
+                    ?: tests.firstOrNull()?.fileName
+            }
+
+            else -> tests.firstOrNull()?.className
+        }
+        val tRun0 = System.currentTimeMillis()
+        val result = testRunner.run(scope)
+        logger.debug("TestingOrchestrator: test run done in ${System.currentTimeMillis() - tRun0}ms, success=${result.success}")
+
+        // Если фреймворк не найден — предложить инструкции
+        if (!result.success) {
+            val fw = frameworkDetector.detect()
+            if (fw == TestFramework.NONE) {
+                val note = frameworkDetector.suggestAddInstructions(TestFramework.JUNIT5)
+                val augmented = result.reportText + "\n\n" + "Рекомендации по установке тестового фреймворка:\n" + note
+                val total = System.currentTimeMillis() - t0
+                logger.debug("TestingOrchestrator: completed with suggestions in ${total}ms")
+                return result.copy(reportText = augmented)
+            }
+        }
+        val total = System.currentTimeMillis() - t0
+        logger.debug("TestingOrchestrator: completed ok in ${total}ms")
+        return result
+    }
 }
